+++ common/src/main/java/net/conczin/mca/entity/VillagerEntityMCA.java
@@ -636,7 +636,9 @@ public class VillagerEntityMCA extends Villager implements VillagerLike<Villager
         }
 
         if (!level().isClientSide) {
-            if (tickCount % 200 == 0 && getHealth() < getMaxHealth()) {
+            if (tickCount % 200 == 0
+                && getHealth() < getMaxHealth()
+                && getBrain().getMemory(MemoryModuleType.ATTACK_TARGET).isEmpty()) {
                 // if the villager has food they should try to eat.
                 ItemStack food = getMainHandItem();
                 FoodProperties foodProperties = food.get(DataComponents.FOOD);